server file cleanup

+++ src/index.ts
@@ -1,8 +1,12 @@
-import "./env.js"
+import "./env.js" // Import env variables
+
 import express from "express"
 import cors from "cors"
+
+// Prisma
 import { prisma } from "./lib/prisma.js"
-import { clerkMiddleware } from "@clerk/express"
+// Clerk Auth
+import { clerkMiddleware, getAuth } from "@clerk/express"
 
 const app = express();
 
@@ -10,12 +14,33 @@ app.use(cors());
 
 app.use(express.json());
 
+app.use(clerkMiddleware());
+
+// Frontend port default to 5050
+const PORT = process.env.PORT || 5050;
+
+// Unauthorized msg
+const UNAUTHORIZED = `Unauthorized`;
+
+
+/**
+ * RUN BACKEND
+ */
+app.listen(PORT, () => {
+
+  console.log(`Server running on port ${PORT}`);
+  console.log(`Databse URL: ${process.env.DATABASE_URL}`)
+  console.log(`Clerk Publishable Key: ${process.env.CLERK_PUBLISHABLE_KEY}`)
+  console.log(`Clerk Secert Key: ${process.env.CLERK_SECRET_KEY}`)
+});
+
+/**
+ * Endpoint for landing page
+ */
 app.get("/", async (req, res) => {
-
   res.json({
-    message: "Yoripe API running",
+    message: `Yoripe API running`,
   });
-
 });
 
 /**
@@ -29,12 +54,19 @@ app.get("/", async (req, res) => {
  * @return
  */
 app.post("/recipes", async (req, res) => {
+  // Authorization check
+  const { userId } = getAuth(req);
+  if (!userId) {
+    return res.status(401).json({ error: UNAUTHORIZED });
+  }
+
   try {
     // Parse required params from request body
     const { title, description, instructions, ingredients } = req.body;
 
     const newRecipe = await prisma.recipe.create({
       data: {
+        userId,
         title,
         description,
         instructions,
@@ -46,21 +78,28 @@ app.post("/recipes", async (req, res) => {
         ingredients: true,
       },
     });
+
+    const msg = `Created new recipe for user ${userId}`;
+
     // Log recipe created
-    console.log(`Recipe Added: ${newRecipe}`);
+    console.log(msg);
+    console.log(newRecipe);
 
     // 201 response with new recipe
     res.status(201).json({
-      message: "Created new recipe",
+      message: msg,
       newRecipe,
     });
   } catch (error) {
+    const errorMsg = `Failed to create recipe for user ${userId}`;
+
     // Log error
+    console.log(errorMsg);
     console.log(error);
 
     // 500 Error
     res.status(500).json({
-      error: "Failed to create recipe",
+      error: errorMsg,
     });
   }
 });
@@ -77,10 +116,23 @@ app.post("/recipes", async (req, res) => {
  * @returns
  */
 app.get("/recipes", async (req, res) => {
+  // Authorization check
+  const { userId } = getAuth(req);
+  if (!userId) {
+    return res.status(401).json({ error: UNAUTHORIZED });
+  }
+
   const { search, title, ingredient } = req.query;
 
-  const whereAnd: any[] = [];
+  // Default filtering by signed in user
+  const whereAnd: any[] = [
+    {
+      userId,
+    }
+  ];
 
+  // If general search query, filter by given search string by matching to either title,
+  // description or ingredient names
   if (search) {
     whereAnd.push({
       OR: [
@@ -110,6 +162,7 @@ app.get("/recipes", async (req, res) => {
     });
   }
 
+  // If filtering by title
   if (title) {
     whereAnd.push({
       title: {
@@ -119,6 +172,7 @@ app.get("/recipes", async (req, res) => {
     });
   }
 
+  // If filtering by ingrediendts
   if (ingredient) {
     whereAnd.push({
       ingredients: {
@@ -145,18 +199,24 @@ app.get("/recipes", async (req, res) => {
       },
       where: whereAnd.length ? { AND: whereAnd } : undefined,
     });
-    console.log({ msg: "Retrieved all recipes", recipes });
+
+    const msg = `Retrieved all recipes for user ${userId}`;
+
+    console.log({ msg: msg, recipes });
 
     // 200 response and return recipes
     res.status(200).json(recipes);
 
   } catch (error) {
+    const errorMsg = `Failed to retrieve recipes for user ${userId}`;
+
     // Log Error
+    console.log(errorMsg)
     console.log(error);
 
     // 500 Error
     res.status(500).json({
-      error: "Failed to retrieve recipes",
+      error: errorMsg,
     });
   }
 });
@@ -168,26 +228,38 @@ app.get("/recipes", async (req, res) => {
  * @returns
  */
 app.get("/recipes/:id", async (req, res) => {
+  // Authorization check
+  const { userId } = getAuth(req);
+  if (!userId) {
+    return res.status(401).json({ error: UNAUTHORIZED });
+  }
+
   const { id } = req.params;
 
   try {
     const recipe = await prisma.recipe.findUniqueOrThrow({
       where: {
         id: id,
+        userId: userId,
       },
       include: {
         ingredients: true,
       },
     });
 
-    console.log({ msg: `Retreived recipe ${id}`, recipe });
+    const msg = `Retrieved recipe ${id}`;
+
+    console.log({ msg: msg, recipe });
 
     res.status(200).json(recipe);
   } catch (error) {
-    console.log(`Failed to retrieve recipe ${id}`);
+    const errorMsg = `Failed to retrieve recipe ${id}`;
+  
+    console.log(errorMsg);
+    console.log(error);
 
     res.status(500).json({
-      error: `Failed to retrieve recipe ${id}`,
+      error: errorMsg,
     });
   }
 });
@@ -207,6 +279,12 @@ app.get("/recipes/:id", async (req, res) => {
  * @returns
  */
 app.patch("/recipes/:id", async (req, res) => {
+  // Authorization check
+  const { userId } = getAuth(req);
+  if (!userId) {
+    return res.status(401).json({ error: UNAUTHORIZED });
+  }
+
   const { id } = req.params;
   const { title, description, instructions } = req.body;
 
@@ -214,6 +292,7 @@ app.patch("/recipes/:id", async (req, res) => {
     const updatedRecipe = await prisma.recipe.update({
       where: {
         id: id,
+        userId: userId,
       },
       data: {
         title,
@@ -225,17 +304,21 @@ app.patch("/recipes/:id", async (req, res) => {
       },
     });
 
-    console.log(`Updated recipe ${id}`);
+    const msg = `Updated recipe ${id}`;
+    console.log(msg);
 
     res.status(200).json({
-      message: `Updated recipe ${id}`,
+      message: msg,
       updatedRecipe,
     });
   } catch (error) {
+    const errorMsg = `Failed to update recipe ${id}`;
+
+    console.log(errorMsg);
     console.log(error);
 
     res.status(500).json({
-      error: `Failed to update recipe ${id}`,
+      error: errorMsg,
     });
   }
 });
@@ -251,44 +334,41 @@ app.patch("/recipes/:id", async (req, res) => {
  * @returns
  */
 app.delete("/recipes/:id", async (req, res) => {
+  // Authorization check
+  const { userId } = getAuth(req);
+  if (!userId) {
+    return res.status(401).json({ error: UNAUTHORIZED });
+  }
+  
   const { id } = req.params;
 
   try {
 
     const deletedRecipe = await prisma.recipe.delete({
       where: {
-        id: id
+        id: id,
+        userId: userId,
       },
     });
+    const msg = `Successfully deleted recipe ${id}`
 
-    console.log(`Deleted recipe ${id}`);
+    console.log(msg);
 
     // 200 to return deleted recipe
     res.status(200).json({
-      message: `Successfully deleted recipe ${id}`,
+      message: msg,
       deletedRecipe,
     })
   } catch (error) {
+    const errorMsg = `Failed to delete recipe ${id}`;
+
     // Log error
+    console.log(errorMsg);
     console.log(error);
 
     // 500 error
     res.status(500).json({
-      error: `Failed to delete recipe ${id}`,
+      error: errorMsg,
     });
   }
-});
-
-
-/**
- * RUN BACKEND
- */
-
-const PORT = process.env.PORT || 5050;
-
-app.listen(PORT, () => {
-
-  console.log(`Server running on port ${PORT}`);
-  console.log(`Databse URL: ${process.env.DATABASE_URL}`)
-
 });
